+++ public/index.html
@@ -440,7 +440,7 @@
       </article>
     </template>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
 
   </body>
 </html>
