+++ src/main.js
@@ -1,4 +1,5 @@
-'use strict';
+import makeFilter from './make-filter.js';
+import {generateTask} from "./make-task";
 
 const MAX_TASK_COUNT = 10;
 const MIN_TASK_COUNT = 1;
@@ -6,39 +7,6 @@ const TASK_NUMBER = 7;
 
 const boardTask = document.querySelector(`.board__tasks`);
 const filterSection = document.querySelector(`.main__filter`);
-
-// функция отрисовки фильтра
-/**
- *
- * @param {string} title
- * @param {number} amount
- * @param {boolean} isChecked
- * @return {string}
- */
-const getElementFilter = function (title, amount, isChecked = false) {
-  return `
-    <input
-      type="radio"
-      id="filter__${title.toLowerCase()}"
-      class="filter__input visually-hidden"
-      name="filter"
-      ${isChecked ? ` checked` : ``}
-      ${amount === 0 ? ` disabled` : `` }
-      />
-    <label for="filter__${title.toLowerCase()}" class="filter__label">
-    ${title}<span class="filter__${title.toLowerCase()}-count"> ${amount}</span></label
-    >
-    `;
-};
-
-// функция отрисовки карточки задания
-/**
- * @param {Node} template
- * @return {Node}
- */
-const generateTask = function (template) {
-  return document.importNode(template.content, true);
-};
 
 // функция создания борда карточек
 /**
@@ -57,13 +25,13 @@ const generateTaskList = function (board, count) {
   board.appendChild(baseFragment);
 };
 
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`All`, 15, true));
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`Overdue`, 0));
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`Today`, 0));
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`Favorites`, 7));
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`Repeating`, 2));
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`Tags`, 6));
-filterSection.insertAdjacentHTML(`beforeend`, getElementFilter(`Archive`, 115));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`All`, 15, true));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`Overdue`, 0));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`Today`, 0));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`Favorites`, 7));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`Repeating`, 2));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`Tags`, 6));
+filterSection.insertAdjacentHTML(`beforeend`, makeFilter(`Archive`, 115));
 
 generateTaskList(boardTask, TASK_NUMBER);
